+++ Assets/MonsterPack3D/Animations/Bat/SA_Bat_Bite.anim
@@ -14158,4 +14158,11 @@ AnimationClip:
     flags: 0
   m_HasGenericRootTransform: 0
   m_HasMotionFloatCurves: 0
-  m_Events: []
+  m_Events:
+  - time: 0.25
+    functionName: PerformAttack
+    data: 
+    objectReferenceParameter: {fileID: 2096101378255670412, guid: 55f0b9b41e413174592474649b069fd5, type: 3}
+    floatParameter: 0
+    intParameter: 0
+    messageOptions: 0
