+++ Assets/MonsterPack3D/Animations/Mushroom/SA_Mushroom_Headbutt.anim
@@ -445,8 +445,7 @@ AnimationClip:
         time: 1.0416667
         value: {x: 0.00050843466, y: -0.000011994608, z: -0.0000013719037, w: 0.9999999}
         inSlope: {x: 0.000000021653271, y: 0.00000003735844, z: 0.0000012509402, w: 0}
-        outSlope: {x: 0.000000021653271, y: 0.00000003735844, z: 0.0000012509402,
-          w: 0}
+        outSlope: {x: 0.000000021653271, y: 0.00000003735844, z: 0.0000012509402, w: 0}
         tangentMode: 0
         weightedMode: 0
         inWeight: {x: 0.33333334, y: 0.33333334, z: 0.33333334, w: 0.33333334}
@@ -1728,10 +1727,8 @@ AnimationClip:
       - serializedVersion: 3
         time: 1.375
         value: {x: -0.0005101851, y: 0.000012005065, z: -0.0000011374871, w: 0.9999999}
-        inSlope: {x: -0.000000045402018, y: -0.000000087169695, z: 0.0000035505457,
-          w: 0}
-        outSlope: {x: -0.000000045402018, y: -0.000000087169695, z: 0.0000035505457,
-          w: 0}
+        inSlope: {x: -0.000000045402018, y: -0.000000087169695, z: 0.0000035505457, w: 0}
+        outSlope: {x: -0.000000045402018, y: -0.000000087169695, z: 0.0000035505457, w: 0}
         tangentMode: 0
         weightedMode: 0
         inWeight: {x: 0.33333334, y: 0.33333334, z: 0.33333334, w: 0.33333334}
@@ -1739,10 +1736,8 @@ AnimationClip:
       - serializedVersion: 3
         time: 1.7916667
         value: {x: -0.0005101851, y: 0.000012005065, z: -0.0000011374872, w: 0.9999999}
-        inSlope: {x: -0.000000006984926, y: -0.000000011405074, z: 0.00000034415086,
-          w: 0}
-        outSlope: {x: -0.000000006984926, y: -0.000000011405074, z: 0.00000034415086,
-          w: 0}
+        inSlope: {x: -0.000000006984926, y: -0.000000011405074, z: 0.00000034415086, w: 0}
+        outSlope: {x: -0.000000006984926, y: -0.000000011405074, z: 0.00000034415086, w: 0}
         tangentMode: 0
         weightedMode: 0
         inWeight: {x: 0.33333334, y: 0.33333334, z: 0.33333334, w: 0.33333334}
@@ -26231,4 +26226,11 @@ AnimationClip:
     flags: 0
   m_HasGenericRootTransform: 0
   m_HasMotionFloatCurves: 0
-  m_Events: []
+  m_Events:
+  - time: 1.4583334
+    functionName: PerformAttack
+    data: 
+    objectReferenceParameter: {fileID: 7369327913803325875, guid: e8e7410d68db893459d8d0d004f05576, type: 3}
+    floatParameter: 0
+    intParameter: 0
+    messageOptions: 0
