+++ Assets/_Assets/UI/UIComponentPrefab/IconBasePrefab.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: e6ba42578b13a574f83603acde6bb78d
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
