--- Assets/_Assets/UI/UIPrefab.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 0666e19e1d7f4f748935ca8e69319f79
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
