--- Assets/_Assets/UI/UIPrefab/ActionBarSlot.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 4cd70f5eb6e87ac48877a929716181a2
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
