--- Assets/_Assets/UI/UIPrefab/BuffBaseIcon.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: c5ed1201b5d15474eb77aa7b5ce9f1cc
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
