--- Assets/_Assets/UI/UIPrefab/DamageText.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: d3ee393ac8ad9c64c9590f0f0977bf96
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
