--- Assets/_Assets/UI/UIPrefab/InventorySlot.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 1d9cc13819a3424499eb68002fe1d758
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
