--- Assets/_Assets/_Anim/AnimationController/Male_Playercotroller.controller.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: d2d7419deee50364582f0956fe9a4f30
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 9100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
