+++ Assets/_Assets/_Font.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7057bae6f07d90f43ba1fcb2c30a8c91
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
