+++ Assets/_Assets/_Font/NotoSansKR-Bold SDF.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4176c19ffd5d678429722c1eceb4e68e
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
