+++ Assets/_Script/MonsterSC/OBJ_SC/MushRoomData.asset
@@ -15,7 +15,7 @@ MonoBehaviour:
   monsterName: MushRoomMon
   monsterType: 0
   HP: 800
-  moveSpeed: 5
+  moveSpeed: 3
   movableArea: 50
   movingAreaRedius: 30
   detectionRadius: 15
