+++ Assets/_Script/UIControlSC/MonsterUI/MonsterDamage.cs
@@ -14,14 +14,15 @@ public class MonsterDamage : MonoBehaviour
         timer = 0.0f;
         transform.localPosition = Vector3.zero;
         transform.localEulerAngles = Vector3.zero;
+        damageText.GetComponentInChildren<TextMeshPro>();
     }
-
 
     private void Update()
     {
         transform.position += Vector3.up * speed * Time.deltaTime;
 
         timer += Time.deltaTime;
+        FadeOut();
         if (timer >= lifeTime)
         {
             Destroy(gameObject);
@@ -31,5 +32,9 @@ public class MonsterDamage : MonoBehaviour
     public void SetDamage(int damage)
     {
         GetComponent<TextMeshPro>().text = damage.ToString();
+    }
+
+    void FadeOut(){
+        damageText.alpha = Mathf.Lerp(1f,0f,timer/lifeTime);
     }
 }
